Passando a função de calcula pageRank para atleticaRepositoryImpl
